requirements.txt
fastapi
sentence-transformers
pypdf
scikit-learn
numpy
python-dotenv
requests
faiss-cpu
python-multipart